add spec

+++ app/controllers/authentications_controller.rb
@@ -2,8 +2,11 @@ class AuthenticationsController < ApplicationController
   before_filter :login_required, :only => %w(destroy)
 
   def destroy
-    authentication = user.authentications.where(:id => params[:id]).last
-    authentication.destroy
+    if user.multi_account?
+      authentication = user.authentications.where(:id => params[:id]).last
+      authentication.destroy
+    end
+
     redirect_to edit_account_path
   end
 end
